Reprint Server: Separate the plugin and compatibility test suites
Splits SiteExportSecretTest into a canonical ReprintServerPluginTest and
a deletable ReprintServerCompatTest, sharing a harness that boots
through the namespaced constants instead of compat.php's SITE_EXPORT_*
aliases. Adds three tests covering the released constant map, released
function delegation, and legacy constant adoption, which nothing covered
before.
